usuario y libro backend

+++ src/main/java/com/biblioteca/repository/LibroRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface LibroRepository extends JpaRepository<Libro, Long> {
+public interface LibroRepository extends JpaRepository<Libro, Integer> {
     Optional<Libro> findByIsbn(String isbn);
 }
